+++ tomcat/src/main/java/org/apache/coyote/http11/body/EmptyBody.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.body;
 
 public class EmptyBody implements Body {
     @Override
