+++ tomcat/src/main/java/org/apache/coyote/http11/body/FormUrlEncodedBody.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.body;
 
 import java.util.HashMap;
 import java.util.List;
