+++ tomcat/src/main/java/org/apache/coyote/http11/header/HeaderType.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.header;
 
 public enum HeaderType {
     CONTENT_TYPE("Content-Type"),
