+++ tomcat/src/main/java/org/apache/coyote/http11/header/Headers.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.header;
 
 import java.util.EnumMap;
 import java.util.List;
