+++ tomcat/src/main/java/org/apache/coyote/http11/header/HttpMethod.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.header;
 
 public enum HttpMethod {
     GET,
