+++ tomcat/src/main/java/org/apache/coyote/http11/header/HttpVersion.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.header;
 
 public enum HttpVersion {
     HTTP_1_1("HTTP/1.1"),
