+++ tomcat/src/main/java/org/apache/coyote/http11/header/RequestLine.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.header;
 
 import java.net.URI;
